+++ lib/compendium-api.rb
@@ -1,13 +1,15 @@
+# Base Class
 require 'nretnil/compendium-api/compendium'
-
+# Endpoint Classes
 require 'nretnil/compendium-api/calendar'
 require 'nretnil/compendium-api/category'
 require 'nretnil/compendium-api/comment'
 require 'nretnil/compendium-api/content'
 require 'nretnil/compendium-api/content_group'
+require 'nretnil/compendium-api/custom_field'
 require 'nretnil/compendium-api/export'
 require 'nretnil/compendium-api/publisher'
 require 'nretnil/compendium-api/user'
 require 'nretnil/compendium-api/version'
-
+# Helper Class
 require 'nretnil/compendium-helpers/helpers.rb'